Add skeletons for the methods involved in the two key algorithms
CondList is the algorithm which, given a tree and a variable ID,
determines the list of variables to condition on. CVI refers to the
entire algorithm, which, given a tree and a variable ID, uses CondList
to compute the variables to condition on, determines a valid
permutation, and computes the mean decrease in accuracy.